Move unit tests for MergeColumns into simd/ut folder
Also:
- Move unit tests for simd library into ut folder so that the tests
are in one place.
- Add comment and debug output into ut for MergeColumns.
- Comment some tests that cause infinite wait (they should pass).
