--- ydb/library/yql/utils/simd/exec/runtime_dispatching/ut/ya.make
@@ -1,7 +1,0 @@
-UNITTEST_FOR(ydb/library/yql/utils/simd/exec/runtime_dispatching)
-
-SRCS(
-    ut.cpp
-)
-
-END()
