+++ ydb/library/yql/utils/simd/exec/runtime_dispatching/ya.make
@@ -16,7 +16,3 @@ RECURSE(
     Fallback_algo
     main
 )
-
-RECURSE_FOR_TESTS(
-    ut
-)
