+++ ydb/library/yql/utils/simd/ut/ya.make
@@ -2,6 +2,7 @@ UNITTEST_FOR(ydb/library/yql/utils/simd)
 
 SRCS(
     simd_ut.cpp
+    merge_ut.cpp
 )
 
 END()
